Updated my fragment methods to allow overlay for add more  add more now shows a dialog with a list of containers
i added an xml file for the overlay dialog_maintanance_catergories
  intention is to dynaically grab all the tables Sid created with perpolated data and present to user for selection

tested code to make sure all older functionality is working will push to prod but will follow up with team to review feedback as the spinner is now obsolete. maybe we add animation to the bottom portion same as the main page dependant on active vehicle.